Add files via upload
The main idea is that a sorting algorithm is applied on a array, row by row. Each step of the sorting process
is stored, and an animated visual is then created. By using different sorting algorithms, one can then produce
different visuals that showcase the features and specificities of each method